+++ transport/internet/tls/acme.go
@@ -3,11 +3,13 @@ package tls
 import (
 	"context"
 	"crypto/tls"
+	"strings"
 
 	"github.com/caddyserver/certmagic"
 	"github.com/libdns/cloudflare"
 )
 
+// Start an ACME service
 func StartACME(ACMEToken string, ACMEMail string, ServerName string) *certmagic.Config {
 	if ACMEMail == "" {
 		ACMEMail = "love@xray.com" // ¿
@@ -38,7 +40,6 @@ func StartACME(ACMEToken string, ACMEMail string, ServerName string) *certmagic.
 	})
 	config = certmagic.New(cache, *config)
 
-	// 管理证书 TODO
 	err := config.ManageAsync(context.Background(), []string{ServerName})
 	if err != nil {
 		return nil
@@ -47,10 +48,51 @@ func StartACME(ACMEToken string, ACMEMail string, ServerName string) *certmagic.
 	return config
 }
 
+// Get certificat
 func GetACMECertificate(ACMEService *certmagic.Config, hello *tls.ClientHelloInfo) *tls.Certificate {
 	cert, err := ACMEService.GetCertificate(hello)
 	if err != nil {
 		return nil
 	}
 	return cert
+}
+
+
+// An Option to change CertificateFunc to GetNewGetACMECertificateFunc()
+func WithACME(rejectUnknownSNI bool, ACMEService *certmagic.Config) Option {
+	return func(config *tls.Config) {
+		config.GetCertificate = GetNewGetACMECertificateFunc(rejectUnknownSNI, ACMEService)
+	}
+}
+
+func GetNewGetACMECertificateFunc(rejectUnknownSNI bool, ACMEService *certmagic.Config) func(hello *tls.ClientHelloInfo) (*tls.Certificate, error) {
+	return func(hello *tls.ClientHelloInfo) (*tls.Certificate, error) {
+		var certs []*tls.Certificate
+		sni := strings.ToLower(hello.ServerName)
+		certs = append(certs, GetACMECertificate(ACMEService, hello))
+		if len(certs) == 0 {
+			return nil, errNoCertificates
+		}
+		if !rejectUnknownSNI && (len(certs) == 1 || sni == "") {
+			return certs[0], nil
+		}
+		gsni := "*"
+		if index := strings.IndexByte(sni, '.'); index != -1 {
+			gsni += sni[index:]
+		}
+		for _, keyPair := range certs {
+			if keyPair.Leaf.Subject.CommonName == sni || keyPair.Leaf.Subject.CommonName == gsni {
+				return keyPair, nil
+			}
+			for _, name := range keyPair.Leaf.DNSNames {
+				if name == sni || name == gsni {
+					return keyPair, nil
+				}
+			}
+		}
+		if rejectUnknownSNI {
+			return nil, errNoCertificates
+		}
+		return certs[0], nil
+	}
 }
